requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
typer==0.9.0
rich==13.7.0
pydantic==2.6.4
rank-bm25==0.2.2
numpy==1.26.4
pyyaml==6.0.1
pytest==8.1.1
mypy==1.9.0
flake8==7.0.0
black==24.2.0
isort==5.13.2
